+++ dags/dag_owners.py
@@ -20,8 +20,8 @@ owners = {
     "miovision_pull": ["natalie", "gabe"],
     "monitor_db_size": ["gabe", "natalie"],
     "open_data_checks": ["gabe"],
-    "pull_here": ["natalie"],
-    "pull_here_path": ["natalie"],
+    "pull_here": ["natalie", "gabe"],
+    "pull_here_path": ["natalie", "gabe"],
     "pull_here_path_hm": ["natalie", "gabe"],
     "queens_park_aggregation": ["natalie"],
     "rapidto_aggregation": ["natalie"],
